validaciones y documentacion

+++ microservice-transactions/src/main/java/com/ntt/microservicetransactions/domain/model/dto/UpdatedBankAccountDTO.java
@@ -3,15 +3,17 @@ package com.ntt.microservicetransactions.domain.model.dto;
 import lombok.Getter;
 import lombok.Setter;
 
+/**
+ * Represents data transfer object of updated bank account
+ */
 @Getter
 @Setter
 public class UpdatedBankAccountDTO {
-
     private String accountNumber;
     private boolean completedTransaction;
-    private double amount;
+    private float amount;
     private String typeTransaction;
-    private double balance;
+    private float balance;
     private int completedTransactions;
 
 }
